Separating data from the services and projects section.
