+++ frontend/src/shared/dbImages.ts
@@ -1,0 +1,52 @@
+//residencial images-cards
+import projImg1 from "../assets/img/recidencial-fachada-moderna-esquina-espelhada-jayarquiteta.webp";
+import projImg2 from "../assets/img/recidencial-fachada-moderna-jayarquiteta.webp";
+import projImg3 from "../assets/img/recidencial-fachada-moderna-planta-jayarquiteta.webp";
+import projImg4 from "../assets/img/residencial-fachada-muro-moderna-jayarquiteta.webp";
+import projImg5 from "../assets/img/residencial-fachada-residencial-moderna-jayarquiteta.webp";
+import projImg6 from "../assets/img/residencial-fachada-residencial-jayarquiteta-muro.webp";
+
+//interiores images-cards
+import interiorImg1 from "../assets/img/interiores-quarto-moderno-suite-planejado-clean-jayarquiteta.webp";
+import interiorImg2 from "../assets/img/interiores-homeoffice-escritorio-jayarquiteta.webp";
+import interiorImg3 from "../assets/img/interiores-cozinha-moderna-preta-planejada-jayarquiteta.webp";
+import interiorImg4 from "../assets/img/interiores-quarto-casal-suite-planejada-moderna-jayarquiteta.webp";
+import interiorImg5 from "../assets/img/interiores-saladetv-moderna-preta-pequena-jayarquiteta.webp";
+import interiorImg6 from "../assets/img/interiores-areagourmet-preto-externo-indrustrial-jayarquiteta.webp";
+
+//comercial images-cards
+import comercialImg1 from "../assets/img/comercial-fachada-loja-moderna-arquitetura.webp";
+import comercialImg2 from "../assets/img/comercial-fachada-moderna-jayarquiteta.webp";
+import comercialImg3 from "../assets/img/comercial-clinica-projetos-planejada-moderna-jayarquiteta.webp";
+import comercialImg4 from "../assets/img/comercial-clinica-moderna-clean-branca-jayarquiteta.webp";
+import comercialImg5 from "../assets/img/comercial-clinica-pediatria-moderna-tematica-jayarquiteta.webp";
+import comercialImg6 from "../assets/img/comercial-escritorio-industrial-moderno-jayarquiteta.webp";
+
+// serviços icons-images-card
+import construcaoImg1 from "../assets/img/contrucao-projeto.png";
+import reformaImg2 from "../assets/img/reforma-arquitetonico-residencia.png";
+import interioresImg3 from "../assets/img/interiores-projetos-decoracao.png";
+
+export default {
+  projImg1,
+  projImg2,
+  projImg3,
+  projImg4,
+  projImg5,
+  projImg6,
+  interiorImg1,
+  interiorImg2,
+  interiorImg3 ,
+  interiorImg4 ,
+  interiorImg5 ,
+  interiorImg6 ,
+  comercialImg1,
+  comercialImg2,
+  comercialImg3,
+  comercialImg4,
+  comercialImg5,
+  comercialImg6,
+  construcaoImg1,
+  reformaImg2,
+  interioresImg3
+}
